Ignore all latedefs in versionmath.js
There are many instances in this file where a variable or function is
defined after it is used.  Rather than tell JSHint explicitly each
time, it is simpler to ignore all latedefs in the whole file.

This also avoids having to refactor the entire file to fix this type
of linting error.  Such a refactor introduces an unnecessary
regression risk.